changed int to double in SumOfPrice in GarageOverviewViewModel. Commented out SumOfPrice  in Overview() in GarageController

+++ GarageMVC/GarageMVC/ViewModels/GarageOverviewViewModel.cs
@@ -5,6 +5,6 @@ namespace GarageMVC.ViewModels
     {
         public int SumOfAllWheels { get; internal set; }
         public List<string> VehicleType { get; internal set; }
-        public int SumOfPrice { get; internal set; }
+        public double SumOfPrice { get; internal set; }
     }
 }
